Backend/requirements.txt
Flask
PyMySQL
python-dotenv
gunicorn
flask-cors